feat: switch referee exams to absolute point scoring
Replace percentage-based pass/fail with absolute point thresholds that
support F-course L1/L2/L3 score bands and G-course pass/fail (L3).

- Questions and exam question snapshots carry an explicit integer point
  value (1, 2, or 3); snapshot creation copies the value from the source
  question.
- Scoring awards full points or zero per question — no partial-credit path
  remains active for online referee exams.
- course_type_question_distributions and exams replace pass_percentage with
  l1_threshold, l2_threshold, l3_threshold (F-course) and pass_threshold
  (G-course), snapshotted at exam creation time.
- Scoring.compute_total/2 now returns %{achieved_points, max_points, outcome}
  where outcome is one of :l3_pass, :l2_pass, :l1_eligible, :fail, or
  :not_applicable (J-course guard).
- exam_participants gain achieved_points, max_points, and exam_outcome fields
  alongside the kept score/max_score/passed legacy columns.
- Tests updated: partial-credit tests removed, new tests cover exact-match
  scoring, mixed point totals, all F-course and G-course threshold boundaries,
  and unanswered questions.

+++ lib/whistle/exams/exam_answer.ex
@@ -10,7 +10,7 @@ defmodule Whistle.Exams.ExamAnswer do
     field :question_type, :string
     field :text_answer, :string
     field :is_correct, :boolean
-    field :awarded_points, :decimal
+    field :awarded_points, :integer
     field :answered_at, :naive_datetime
 
     has_many :answer_choices, Whistle.Exams.ExamAnswerChoice, foreign_key: :exam_answer_id
